+++ actions/playerMatchesActions.js
@@ -1,5 +1,6 @@
 import fetch from 'isomorphic-fetch';
 import { HOST_URL } from './';
+import { playerMatchTransform } from '../transformations';
 
 const url = (playerId, numMatches) => `/api/players/${playerId}/matches?limit=${numMatches}`;
 
@@ -33,43 +34,6 @@ const getPlayerMatchesError = (payload) => ({
   type: ERROR,
   payload,
 });
-
-const playerMatchTransformSwitch = (match, field, heroes) => {
-  switch (field) {
-    case 'hero_id':
-      return `${HOST_URL}${heroes[match[field]].img}`;
-    case 'radiant_win':
-      return match.player_slot < 64 && match[field] || match.player_slot >= 64 && !match[field] ? 'W' : 'L';
-    case 'game_mode':
-      return match[field] === 22 ? 'Ranked All Pick' : 'All Pick';
-    case 'start_time': {
-      const today = new Date();
-      const matchDate = new Date(match[field] * 1000);
-      const leftoverDays = Math.floor((today - matchDate) / 1000 / 60 / 60 / 24);
-      switch (leftoverDays) {
-        case 0:
-          return 'Today';
-        case 1:
-          return 'Yesterday';
-        default:
-          return `${leftoverDays} days ago`;
-      }
-    }
-    case 'duration':
-      return `${Math.floor(match[field] / 60)}:${match[field] % 60}`;
-    default:
-      return match[field];
-  }
-};
-
-// TODO - memoize this with selectors
-const playerMatchTransform = (match) => (heroes) => {
-  const transformedMatch = { ...match };
-  Object.keys(match).forEach((field) => {
-    transformedMatch[`${field}Display`] = playerMatchTransformSwitch(match, field, heroes);
-  });
-  return transformedMatch;
-};
 
 export const getPlayerMatches = (playerId, numMatches, host = HOST_URL) => (dispatch, getState) => {
   dispatch(getPlayerMatchesRequest());
